+++ src/data/usecases/survey-result/load-survey-result/db-load-survey-result.ts
@@ -1,6 +1,4 @@
-import { LoadSurveyResultRepository } from '@/data/protocols/db/survey-result/load-survey-result-repository'
-import { LoadSurveyResult } from '@/domain/usecases/survey-result/load-survey-result'
-import { SurveyResultModel } from '../save-survey-result/db-save-survey-result-protocols'
+import { LoadSurveyResult, LoadSurveyResultRepository, SurveyResultModel } from './db-load-survey-result-protocols'
 
 export class DbLoadSurveyResult implements LoadSurveyResult {
   constructor (private readonly loadSurveyResultRepository: LoadSurveyResultRepository) {}
